organize types: PredefinedPackage
