+++ types/PredefinedPackage/DHLExpress.d.ts
@@ -1,0 +1,15 @@
+export declare type PredefinedPackageDHLExpress =
+  | 'JumboDocument'
+  | 'JumboParcel'
+  | 'Document'
+  | 'DHLFlyer'
+  | 'Domestic'
+  | 'ExpressDocument'
+  | 'DHLExpressEnvelope'
+  | 'JumboBox'
+  | 'JumboJuniorDocument'
+  | 'JuniorJumboBox'
+  | 'JumboJuniorParcel'
+  | 'OtherDHLPackaging'
+  | 'Parcel'
+  | 'YourPackaging';
