+++ types/PredefinedPackage/DPDUK.d.ts
@@ -1,0 +1,6 @@
+export declare type PredefinedPackageDPDUK =
+  | 'Parcel'
+  | 'Pallet'
+  | 'ExpressPak'
+  | 'FreightParcel'
+  | 'Freight';
