+++ types/PredefinedPackage/Estafeta.d.ts
@@ -1,0 +1,1 @@
+export declare type PredefinedPackageEstafeta = 'ENVELOPE' | 'PARCEL';
