+++ types/PredefinedPackage/Fastway.d.ts
@@ -1,0 +1,1 @@
+export declare type PredefinedPackageFastway = 'Parcel' | 'Satchel';
