+++ types/PredefinedPackage/FedEx.d.ts
@@ -1,0 +1,11 @@
+export declare type PredefinedPackageFedEx =
+  | 'FedExEnvelope'
+  | 'FedExBox'
+  | 'FedExPak'
+  | 'FedExTube'
+  | 'FedEx10kgBox'
+  | 'FedEx25kgBox'
+  | 'FedExSmallBox'
+  | 'FedExMediumBox'
+  | 'FedExLargeBox'
+  | 'FedExExtraLargeBox';
