+++ types/PredefinedPackage/InterlinkExpress.d.ts
@@ -1,0 +1,6 @@
+export declare type PredefinedPackageInterlinkExpress =
+  | 'Parcel'
+  | 'Pallet'
+  | 'ExpressPak'
+  | 'FreightParcel'
+  | 'Freight';
