+++ types/PredefinedPackage/LaserShipV2.d.ts
@@ -1,0 +1,1 @@
+export declare type PredefinedPackageLaserShipV2 = 'Envelope' | 'Custom';
