+++ types/PredefinedPackage/Liefery.d.ts
@@ -1,0 +1,5 @@
+export declare type PredefinedPackageLiefery =
+  | 'SmallParcel'
+  | 'MediumParcel'
+  | 'LargeParcel'
+  | 'ExtraLargeParcel';
