+++ types/PredefinedPackage/OmniParcel.d.ts
@@ -1,0 +1,11 @@
+export declare type PredefinedPackageOmniParcel =
+  | 'Bag'
+  | 'Box'
+  | 'Carton'
+  | 'Container'
+  | 'Crate'
+  | 'Envelope'
+  | 'Pail'
+  | 'Pallet'
+  | 'Satchel'
+  | 'Tub';
