+++ types/PredefinedPackage/OnTrac.d.ts
@@ -1,0 +1,1 @@
+export declare type PredefinedPackageOnTrac = 'Letter';
