+++ types/PredefinedPackage/Purolator.d.ts
@@ -1,0 +1,5 @@
+export declare type PredefinedPackagePurolator =
+  | 'CustomerPackaging'
+  | 'ExpressPack'
+  | 'ExpressBox'
+  | 'ExpressEnvelope';
