+++ types/PredefinedPackage/RoyalMail.d.ts
@@ -1,0 +1,6 @@
+export declare type PredefinedPackageRoyalMail =
+  | 'Letter'
+  | 'LargeLetter'
+  | 'SmallParcel'
+  | 'MediumParcel'
+  | 'Parcel (for use with RoyalMail24 or RoyalMail48)';
