+++ types/PredefinedPackage/SprintShip.d.ts
@@ -1,0 +1,6 @@
+export declare type PredefinedPackageSprintShip =
+  | 'ENVELOPE'
+  | 'LARGE_BOX_OR_TOTE'
+  | 'MULTI_PACK'
+  | 'SMALL_OR_MEDIUM_BOX_OR_TOTE'
+  | 'PACKAGE';
