+++ types/PredefinedPackage/StarTrack.d.ts
@@ -1,0 +1,9 @@
+export declare type PredefinedPackageStarTrack =
+  | 'Carton'
+  | 'Pallet'
+  | 'Satchel'
+  | 'Bag'
+  | 'Envelope'
+  | 'Item'
+  | 'Jiffybag'
+  | 'Skid';
