+++ types/PredefinedPackage/TForce.d.ts
@@ -1,0 +1,1 @@
+export declare type PredefinedPackageTForce = 'Parcel' | 'Letter';
