+++ types/PredefinedPackage/UPS.d.ts
@@ -1,0 +1,10 @@
+export declare type PredefinedPackageUPS =
+  | 'UPSLetter'
+  | 'UPSExpressBox'
+  | 'UPS25kgBox'
+  | 'UPS10kgBox'
+  | 'Tube'
+  | 'Pak'
+  | 'SmallExpressBox'
+  | 'MediumExpressBox'
+  | 'LargeExpressBox';
