+++ types/PredefinedPackage/USPS.d.ts
@@ -1,0 +1,27 @@
+export declare type PredefinedPackageUSPS =
+  | 'Card'
+  | 'Letter'
+  | 'Flat'
+  | 'FlatRateEnvelope'
+  | 'FlatRateLegalEnvelope'
+  | 'FlatRatePaddedEnvelope'
+  | 'FlatRateGiftCardEnvelope'
+  | 'FlatRateWindowEnvelope'
+  | 'FlatRateCardboardEnvelope'
+  | 'SmallFlatRateEnvelope'
+  | 'Parcel'
+  | 'LargeParcel'
+  | 'IrregularParcel'
+  | 'SoftPack'
+  | 'SmallFlatRateBox'
+  | 'MediumFlatRateBox'
+  | 'LargeFlatRateBox'
+  | 'LargeFlatRateBoxAPOFPO'
+  | 'LargeFlatRateBoardGameBox'
+  | 'RegionalRateBoxA'
+  | 'RegionalRateBoxB'
+  | 'FlatTubTrayBox'
+  | 'EMMTrayBox'
+  | 'FullTrayBox'
+  | 'HalfTrayBox'
+  | 'PMODSack';
