+++ types/PredefinedPackage/index.d.ts
@@ -1,0 +1,36 @@
+import { PredefinedPackageDHLExpress } from './DHLExpress';
+import { PredefinedPackageDPDUK } from './DPDUK';
+import { PredefinedPackageEstafeta } from './Estafeta';
+import { PredefinedPackageFastway } from './Fastway';
+import { PredefinedPackageFedEx } from './FedEx';
+import { PredefinedPackageInterlinkExpress } from './InterlinkExpress';
+import { PredefinedPackageLaserShipV2 } from './LaserShipV2';
+import { PredefinedPackageLiefery } from './Liefery';
+import { PredefinedPackageOmniParcel } from './OmniParcel';
+import { PredefinedPackageOnTrac } from './OnTrac';
+import { PredefinedPackagePurolator } from './Purolator';
+import { PredefinedPackageRoyalMail } from './RoyalMail';
+import { PredefinedPackageSprintShip } from './SprintShip';
+import { PredefinedPackageStarTrack } from './StarTrack';
+import { PredefinedPackageTForce } from './TForce';
+import { PredefinedPackageUPS } from './UPS';
+import { PredefinedPackageUSPS } from './USPS';
+
+export declare type PredefinedPackage =
+  | PredefinedPackageDHLExpress
+  | PredefinedPackageDPDUK
+  | PredefinedPackageEstafeta
+  | PredefinedPackageFastway
+  | PredefinedPackageFedEx
+  | PredefinedPackageInterlinkExpress
+  | PredefinedPackageLaserShipV2
+  | PredefinedPackageLiefery
+  | PredefinedPackageOnTrac
+  | PredefinedPackagePurolator
+  | PredefinedPackageRoyalMail
+  | PredefinedPackageOmniParcel
+  | PredefinedPackageSprintShip
+  | PredefinedPackageStarTrack
+  | PredefinedPackageTForce
+  | PredefinedPackageUPS
+  | PredefinedPackageUSPS;
