+++ types/index.d.ts
@@ -1,6 +1,7 @@
 import { IBaseObject, IDatedObject, IObjectWithId } from './base';
 import { TErrorCode } from './errors/ErrorCode';
 import { IFieldError } from './errors/FieldError';
+import { PredefinedPackage } from './PredefinedPackage';
 import { ServiceLevel } from './ServiceLevel';
 import { DeepPartial } from './utils';
 
@@ -89,148 +90,6 @@ export declare type Carrier =
   | 'Veho'
   | 'Yanwen'
   | 'Yodel';
-
-export declare type PredefinedPackageDHLExpress =
-  | 'JumboDocument'
-  | 'JumboParcel'
-  | 'Document'
-  | 'DHLFlyer'
-  | 'Domestic'
-  | 'ExpressDocument'
-  | 'DHLExpressEnvelope'
-  | 'JumboBox'
-  | 'JumboJuniorDocument'
-  | 'JuniorJumboBox'
-  | 'JumboJuniorParcel'
-  | 'OtherDHLPackaging'
-  | 'Parcel'
-  | 'YourPackaging';
-export declare type PredefinedPackageDPDUK =
-  | 'Parcel'
-  | 'Pallet'
-  | 'ExpressPak'
-  | 'FreightParcel'
-  | 'Freight';
-export declare type PredefinedPackageEstafeta = 'ENVELOPE' | 'PARCEL';
-export declare type PredefinedPackageFastway = 'Parcel' | 'Satchel';
-export declare type PredefinedPackageFedEx =
-  | 'FedExEnvelope'
-  | 'FedExBox'
-  | 'FedExPak'
-  | 'FedExTube'
-  | 'FedEx10kgBox'
-  | 'FedEx25kgBox'
-  | 'FedExSmallBox'
-  | 'FedExMediumBox'
-  | 'FedExLargeBox'
-  | 'FedExExtraLargeBox';
-export declare type PredefinedPackageInterlinkExpress =
-  | 'Parcel'
-  | 'Pallet'
-  | 'ExpressPak'
-  | 'FreightParcel'
-  | 'Freight';
-export declare type PredefinedPackageLaserShipV2 = 'Envelope' | 'Custom';
-export declare type PredefinedPackageLiefery =
-  | 'SmallParcel'
-  | 'MediumParcel'
-  | 'LargeParcel'
-  | 'ExtraLargeParcel';
-export declare type PredefinedPackageOnTrac = 'Letter';
-export declare type PredefinedPackagePurolator =
-  | 'CustomerPackaging'
-  | 'ExpressPack'
-  | 'ExpressBox'
-  | 'ExpressEnvelope';
-export declare type PredefinedPackageRoyalMail =
-  | 'Letter'
-  | 'LargeLetter'
-  | 'SmallParcel'
-  | 'MediumParcel'
-  | 'Parcel (for use with RoyalMail24 or RoyalMail48)';
-export declare type PredefinedPackageOmniParcel =
-  | 'Bag'
-  | 'Box'
-  | 'Carton'
-  | 'Container'
-  | 'Crate'
-  | 'Envelope'
-  | 'Pail'
-  | 'Pallet'
-  | 'Satchel'
-  | 'Tub';
-export declare type PredefinedPackageSprintShip =
-  | 'ENVELOPE'
-  | 'LARGE_BOX_OR_TOTE'
-  | 'MULTI_PACK'
-  | 'SMALL_OR_MEDIUM_BOX_OR_TOTE'
-  | 'PACKAGE';
-export declare type PredefinedPackageStarTrack =
-  | 'Carton'
-  | 'Pallet'
-  | 'Satchel'
-  | 'Bag'
-  | 'Envelope'
-  | 'Item'
-  | 'Jiffybag'
-  | 'Skid';
-export declare type PredefinedPackageTForce = 'Parcel' | 'Letter';
-export declare type PredefinedPackageUPS =
-  | 'UPSLetter'
-  | 'UPSExpressBox'
-  | 'UPS25kgBox'
-  | 'UPS10kgBox'
-  | 'Tube'
-  | 'Pak'
-  | 'SmallExpressBox'
-  | 'MediumExpressBox'
-  | 'LargeExpressBox';
-export declare type PredefinedPackageUSPS =
-  | 'Card'
-  | 'Letter'
-  | 'Flat'
-  | 'FlatRateEnvelope'
-  | 'FlatRateLegalEnvelope'
-  | 'FlatRatePaddedEnvelope'
-  | 'FlatRateGiftCardEnvelope'
-  | 'FlatRateWindowEnvelope'
-  | 'FlatRateCardboardEnvelope'
-  | 'SmallFlatRateEnvelope'
-  | 'Parcel'
-  | 'LargeParcel'
-  | 'IrregularParcel'
-  | 'SoftPack'
-  | 'SmallFlatRateBox'
-  | 'MediumFlatRateBox'
-  | 'LargeFlatRateBox'
-  | 'LargeFlatRateBoxAPOFPO'
-  | 'LargeFlatRateBoardGameBox'
-  | 'RegionalRateBoxA'
-  | 'RegionalRateBoxB'
-  | 'FlatTubTrayBox'
-  | 'EMMTrayBox'
-  | 'FullTrayBox'
-  | 'HalfTrayBox'
-  | 'PMODSack';
-
-export declare type PredefinedPackage =
-  | PredefinedPackageDHLExpress
-  | PredefinedPackageDPDUK
-  | PredefinedPackageEstafeta
-  | PredefinedPackageFastway
-  | PredefinedPackageFedEx
-  | PredefinedPackageInterlinkExpress
-  | PredefinedPackageLaserShipV2
-  | PredefinedPackageLiefery
-  | PredefinedPackageOnTrac
-  | PredefinedPackagePurolator
-  | PredefinedPackageRoyalMail
-  | PredefinedPackageOmniParcel
-  | PredefinedPackageSprintShip
-  | PredefinedPackageStarTrack
-  | PredefinedPackageTForce
-  | PredefinedPackageUPS
-  | PredefinedPackageUSPS;
 
 /**
  * @see https://www.easypost.com/docs/api/node#verification-object
